--- test/unit/node/test-plugin/.eslintrc.js
@@ -1,9 +1,0 @@
-module.exports = {
-  settings: {
-    "import/resolver": {
-      typescript: {
-        project: __dirname,
-      },
-    },
-  },
-}
